requirements.txt
matrix-nio[e2e]==0.25.2
requests
python-dotenv
